Optimize regular expressions for identifying code types
By only matching one regex with a little bit more complexity and
differentiating the results by a capture group the performance of
checking for matches gets way better. Because this regex has to be
matched for each line it is compiled only once  to save this overhead
for most matches.