Do not repeatedly concatenate lines when parsing diffs
Because string concatenation has to allocate a new string for each
allocation the time required for concatenating `n` lines is `O(n^2)`
(with a constant depending on the line length). This can be optimised to
`O(n)` (this is the purpose of `StringBuilder` after all).

In this case a list of lines is used because this avoids problems with
line ending encoding (which is currently handled by always using "\r\n"
internally). This also makes it easier to insert symbols at the start of
the line (for example plus or minus for diff types).

The use sites of the `DiffNode` label are refactored to prevent
duplicate work (for example joining lines and immediately splitting them again).