Compile regular expressions only once
For performance reasons regexes should be compiled only once, because it
takes considerable time to compile them.

The simplest solution is to compile these regexes in a static context,
so that they are only compiled when the class is loaded. A further
optimisation might be to compile them lazily (only compile regexes when
they are actually needed) to prevent slow startup times.

Because regexes are mostly implementation details they are made private
unless they clearly belong to a public API (e.g. `LINEBREAK_REGEX`).