Only parse the code type when needed
The code type is not needed in all code paths. It needs some time, which
is not negligible (regular expression match), to parse the code type, so
it should not be parsed in all code paths.

Consequently, the variable can be declared in a smaller scope to make
that fact obvious to the reader (and a bit more to the compiler).