Refactor Starfold to adhere to DiffNode's invariants
The time specific children of `DiffNode` do not guarantee any ordering,
but the Starfold transformation requires these nodes to be ordered:
- When `respectNodeOrder` is active.
- For the label of the merged node, which is composed sequentially from
  the merged nodes in the order of the given children.

For that reason Starfold cannot use `DiffNode.getChildren` and has to
use `DiffNode.getAllChildren` instead. Because this list contains both
/before/ and /after/ children the actual computation has to filter the
children to a specific time.

It would also be possible to run the two cases (/before/ and /after/) in
parallel, but this a consumer with the semantic of a coroutine and a
finalisation step to merge the remaining nodes. But this introduces
complexity which doesn't seem adequate.