Use a `BufferedReader` to parse diff trees
By using `BufferedReader`s all the platform dependent line endings are
handled automatically. This also reduces space usage if the input is
read from a file, because not all lines are duplicated in memory at the
same time.

Call sites which generate a `String` and have to use `StringReader` are
subject to future refactoring.