Simplify the contract of `FileUtils.hasExtension`
The new contract doesn't handle the special case of a missing dot in the
given file extension. This is fine, because this method is only called
with constant strings, so no user input has to be changed.

It may have been a better choice to not include the dot in the expected
file extension, but in that case all the call sites would have to be
updated (which results in all file extensions handling functions and
constants being modified), so including the dot is the saner choice in
this code base.

Also the name `hasExtension` doesn't imply any additional file type
checks. So it shouldn't check if a file is a directory (the current use
sites do check this for themselves).