Do not store before and after children in `DiffNode`s
Because the invariant
  `node.getBeforeParent().getBeforeChildren().contains(node)`
holds for each `DiffNode node` the `beforeChildren` and `afterChildren`
lists can be computed and do not have to be stored.

By using this invariant, the check if a node is a before/after child can
be implemented in `O(1)` in contrast to the `O(n)` version which uses
the `beforeChildren` and `afterChildren` lists. This speedup is
especially important for parsing `DiffTree`s.

It is also possible to store `Set`s instead of `List`s for `DiffNode`
children to get a similar time complexity, but as demonstrated by this
commit that memory can be saved. If there is a need to compute a
`beforeChildren` or `afterChildren` set (these lists didn't actually
guarantee any ordering) more efficiently than by traversing
`childOrdering` (which requires about the double time) this would be a
valid approach.