Adapt the line graph exporter to the new `Exporter` interface
Line graph specific export code is moved to the line graph exporter.

This changes the line endings to use the OS dependent representation
through the use of `PrintStream`. Currently there is an unnecessary
conversion from `String` to `ByteArray` and back because `OutputStream`s
are used. This will be optimised in the future by directly outputting to
the desired devices (file and terminal) without converting back to a
`String` representation.