Remove `NodeType.ROOT` in favor of an implicit root
The root node is now implicitly characterised by having no parents (no
after and no before parent). This makes it consistent with some
representations of `DiffTree`s where the root node is printed as an IF
node with the feature mapping `true`.

All representations which did explicitly mark the root node are now
printing an IF node and do not mark it specially. Most notably is the
line graph representation which did encode `NodeType.ROOT` in the node
ID (this is the reason why many IDs in the test cases changed).

For backwards compatibility the `isAnnotation` function of `DiffNode` is
still returning false on root nodes. This is subject to further
refactoring, mostly necessary to handle the difference between
`DiffGraph`s and actual trees.