Use checked exceptions instead of monadic error handling
In Java there is no way to ensure that a result value is used, so errors
might be unhandled (there was at least one case). It also required kind
of a weird programming style which doesn't really fit Java (See the
`AtomicReference<DiffResult<DiffTree>>` for example). Since the
widespread use of lambdas are checked exceptions also a bit problematic,
but Apache commons helps here and makes it acceptable.