Warn if createVariationTree might be given a diff
This heuristic should catch all diffs with modifications and only fail
for very obscurely formatted source code (given the set of languages
which are commonly parsed using DiffDetective). The warning is also
logged for every line which starts with a plus or minus, resulting in
many lines of the same log output if the given file is actually a diff
which should make it easy to detect.

Currently, this can't be disabled, but there might be a demand to
disable it in the future.