Deploy javadoc to GitHub Pages using CI
On every push to `main` the generated javadoc is uploaded as an
artifact. This step is conditional to prevent the generation of
unnecessary artifacts.

A second job, as recommened in the documentation of
`actions/deploy-pages`, will then deploy the javadoc on each push to the
`main` branch to the project specific GitHub Pages. This ensures that
the javadoc is always up-to-date.

Note: GitHub Pages have to be enabled and configured to use deployments
by GitHub Actions. The `main` branch also needs permission to deploy to
the GitHub Pages in the environment settings of `github-pages`.