Refactor the duplication code for `DiffNode`s
There is no need for an instance of `Duplication` so all the methods are
made static. By only traversing the tree once, the code gets shorter and
easier to understand.

Noticed the bug which caused a different order of children in
`childOrder`? That's now fixed, although it didn't seem to be relevant
for the usage of these methods. It was undocumented, though.