Patch Gumtree to add `MappingStore.assertConsistency`
The CompositeMatchers.XyMatcher algorithm adds incompatible matchings
for nodes which already have a matching. As this is the fastest matcher,
it's useful to test it further. Thus, incompatible matchings are ignored
and a debug message is printed.

For example, given a matching between A and B
| src -> dst | dst -> src |
|------------|------------|
| A -> B     | B -> A     |

When the incompatible matching between B and C is added, the result is:
| src -> dst | dst -> src |
|------------|------------|
| A -> B     |            |
| C -> B     | B -> C     |

Thus the matching between B and C is ignored if XyMatcher tries to add
it.