refactor: Use identity for `equals` in nodes
Effectively, the semantic of `isSameAs` and `equals` has been exchanged.

There is a trade off here: We can't just copy nodes arbitrarily anymore
but we gain simpler equality reasoning and some performance.

The `FormalDiffGraph` was originally introduced to simplify comparisons
but that is now no longer the case. If no future use cases emerge we
should consider removing it again.