feat: and refactor: Witness for true-false elimination; DNF-SAT
We started to use the true/false elimination frequently and early for optimizations. This made it hard to remember when and where formulas had already been fixed or not. This commit introduces two things:
First, a witness class for elimination.
Second, a new SAT strategy based on DNFs and a simple heuristic to use it on smal formulas.
Also refactored the SAT class to be more simple.