Add nix build instructions
Usage: Run `nix-build` or `nix build` (currently still an experimental
nix feature) in the top level of the repository. If there are no build
errors, a symlink to the result called `result` is added. It contains a
DiffDetective jar including all dependencies and a small wrapper for
running a main class using
`result/bin/Diffdetective org.variantsync.diffdetective.SomeMainClass`.