fix(docker): extended build and execution scripts to handle new project structure
Docker does not allow access to parent directories during the building of an Image. Thus, we have no access to the required artifacts inside the `replication` directory. To address this problem, the build script now first checks that it is being executed from inside the 'esecfse22' subdir, and, if this is the case, it navigates to the project root directory to start the build project from there.