TODO Use a general tree using VariationLabel for variation trees
Unfortunately, all variation tree specific methods need to be outside of
the TreeNode class because methods of the TreeNode class have to deal
with all labels (the labels don't have to store variation information).
A sub-type seems also impractical because there would be cases where
these two tree would need to be converted at runtime just to access
methods in a VariationNode<L> for a TreeNode<VariationLabel<L>> which
are isomorphic.