test: fix the DiffTreeParser tests for non-GumTree matchers
Previously, all tests would run with the default GumTree matcher instead
of the matcher stated in the test case file name.

As stated in my bachelor thesis, some matchers fail to return a valid
matching. Hence, some test cases need to be removed. In particular, the
theta matcher contains a bug that results in matchings that are
inconsistent (e.g., A->B but B->C). Furthermore, the
gumtree-partition-id matcher tests have been removed. This matcher is
already removed in a newer GumTree version.