refactor: store `endif`s in the label
This is necessary because when `endif`s are fully integrated (included
in deep copies, assertConsistency and comparisons) some assumptions are
broken (which fields are compare by `DiffNode`s) which is especially
annoying in `BadVDiff`. By storing the `endif`s in the label instead of
the nodes directly, no user needs to be updated except if the label is
actually modified (there seems to be no instance of that in this code
base).

This also generalized the concept to trailing lines instead of limiting
the concept to endifs. This is also kind of necessary because not all
labels know the node type.

Depending on the user of `VariationDiffParser`, this might be a breaking
change because we now consider the `#endif` line as part of the label.
Hence, if the `#endif` line changed (e.g., a comment is added) the node
can no longer be `DiffType.NON` but we instead split the node into two
`DiffType.REM` and `DiffType.ADD` nodes. Artifacts which are below such
a split node are thus classified as refactored instead of unchanged.