refactor: refactor VariationUnparser.undiff
This kind of includes a breaking change: Keeping trailing empty lines
and interpreting them as unchanged trailing empty lines. Although
multiple empty lines should never occur in well formatted diffs (because
each line contains at least the diff symbol), this is likely not
intended and thus a bug.

Note that preventing a new line in the special case of an empty output
seems unnecessary but was explicitly done in the old version so the
refactored version does this as well.