fix: use the correct child order in line graph exports
When importing line graphs, we consider the order of the edge lines as
the definite child ordering even if we export the child order using
`ChildOrderEdgeFormat`. However, the previous algorithm blindly exports
all `DiffType.NON` edges disregarding the fact the other edges might
need to be exported first. As the order of edge lines acts as a diff
(before edges act as deletions and after edges as insertions), we need
to compute a line diff of the child orders at both times to know which
edges can be exported as existing at both times (unchanged).

Note that some edges will be exported as inserted and deleted instead of
unmodified although both the child and the parent are the same because
line diffs cannot encode moves. Consumers of line graphs should not
depend on the fact which edges are exported as unchanged because the
diffing algorithm needs to apply heuristics to select these edges.