feat: create an interface for and a collection of differs
Note about the visual abstract image: The image is purposefully placed
in the `src` folder such that it is easy to reference from the Javadoc.
Furthermore, the symlink from `docs` allows easy references from
`README.md` makes it more discoverable. The direction of the symlink
(`docs` to `src` instead of `src` to `docs`) is important because the
Javadoc is moved during the build process. A symlink would thus be
incorrect in the default javadoc location (`target/reports/apidocs/`) or
the final destination (`docs/javadoc`) which is used by the Nix build
and thus for Github pages.